LP-553 Move tests to correct directory
